requirements.txt
flask==2.3.3
pytest==7.4.0
langchain==0.0.277
pypdf==3.15.4
azure-storage-blob==12.17.0
openai==0.28.0
tiktoken==0.4.0
faiss-cpu==1.7.4